feat: navigation flow refractor
