--- app/screens/Home/ConnectToDeviceList/styles.ts
@@ -1,37 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-  headerStyle: {},
-  safeArea: {
-    flex: 1,
-  },
-  scrollView: {
-    flex: 1,
-  },
-  scrollViewContainer: {
-    paddingBottom: 100,
-  },
-  subView: {
-    flex: 1,
-  },
-  noDataButtonsContainer: {
-    marginTop: 20,
-    alignItems: 'center',
-  },
-  fab: {
-    position: 'absolute',
-    right: 16,
-  },
-  cardTablet: {
-    width: '70%',
-    alignSelf: 'center',
-  },
-  navigationButton: {
-    marginHorizontal: 16,
-  },
-});
-
-export default styles;
