+++ app/screens/Home/ManageDevices/styles.ts
