+++ app/screens/Home/MoreTab/styles.ts
