+++ app/screens/Loading/styles.ts
